fix: proxy attribute support improvements and tests
- Add Clone to SolverError (needed for KdlError Debug impl)
- Fix KdlError Debug impl to use miette::Report for global hook
- Remove env var dump from test helper (security concern)
- Add dedicated proxy_opaque.rs test files for JSON and XML
- Move proxy/opaque tests from minimal.rs to dedicated file
- Update format crate matrix with opaque/proxy attribute support
- Create issue #1113 for adding proxy support to yaml/toml/args/urlencoded
